feat: add verbose option
To show the results of the command in real time instead of waiting for
it to end.

feat: add the ignore_return_codes flag

To process well websites whose status code can't be trusted

tests: improve test robustness

Launch the api server once per session instead of on each test and
remove the parallel execution as it was giving errors.

perf: change the webdriver chrome_options to options

As it's going to be deprecated soon

fix: handle redirections well

Before they wouldn't be catched